requirements.txt
fastapi==0.135.1
httpx==0.28.1
langchain_community==0.4.1
langchain_core==1.2.18
langchain_experimental==0.4.1
langchain_ollama==1.0.1
langgraph==1.1.2
nest_asyncio==1.6.0
python-dotenv==1.2.2
PyGithub>=2.0.0
python-telegram-bot==22.6
uvicorn==0.41.0
